pyproject.toml
[tool.poetry]
name = "storm-api"
version = "0.1.0"
description = ""
authors = ["waseem"]
readme = "README.md"
packages = [
    { include = "api" },
    { include = "core" },
    { include = "utils" },
]

[tool.poetry.dependencies]
python =">=3.11,<3.12"

fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
pydantic = "^2.11.4"
httpx = "^0.28.1"
pydantic-settings = "^2.9.1"
wikipedia = "1.4.0"
langchain-text-splitters = "^0.3.8"
trafilatura = "^2.0.0"
langchain-huggingface = "^0.1.2"
diskcache = "^5.6.3"
dspy-ai = "2.5.43"
openai = ">=1.61.1,<1.62.0"
toml = "^0.10.2"
numpy = "1.26.4"
kipp = ">=0.3.2,<0.4.0"
cloudpickle = "^3.1.1"
duckduckgo-search = "^8.0.1"
python-dotenv = "^1.1.0"
google-generativeai = "^0.8.5"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
